[LH-46] Move main activity to app module #8
