+++ app/src/main/java/com/lighthouse/lingo_swap/MainActivity.kt
@@ -1,0 +1,38 @@
+package com.lighthouse.lingo_swap
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupWithNavController
+import com.lighthouse.lingo_swap.databinding.ActivityMainBinding
+import com.lighthouse.navigation.NavigationFlow
+import com.lighthouse.navigation.Navigator
+import com.lighthouse.navigation.ToFlowNavigatable
+import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
+
+@AndroidEntryPoint
+class MainActivity @Inject constructor() : AppCompatActivity(), ToFlowNavigatable {
+    private lateinit var binding: ActivityMainBinding
+
+    @Inject
+    lateinit var navigator: Navigator
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
+
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
+        val navController = navHostFragment.navController
+
+        navigator.navController = navController
+        binding.bottomNav.setupWithNavController(navController)
+
+    }
+
+    override fun navigateToFlow(flow: NavigationFlow) {
+        navigator.navigateToFlow(flow)
+    }
+}
