+++ app/src/main/res/menu/btn_nav.xml
@@ -2,23 +2,24 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <group>
         <item
-            android:id="@+id/item_home"
+            android:id="@id/home_flow"
             android:icon="@drawable/home"
             android:title="@string/btn_nav_home" />
 
         <item
-            android:id="@+id/item_chat"
+            android:id="@id/chats_flow"
             android:icon="@drawable/chat_icon"
             android:title="@string/btn_nav_chat" />
 
         <item
-            android:id="@+id/item_board"
+            android:id="@id/board_flow"
             android:icon="@drawable/board_icon"
             android:title="@string/btn_nav_board" />
 
         <item
-            android:id="@+id/item_profile"
+            android:id="@id/profile_flow"
             android:icon="@drawable/profile_icon"
             android:title="@string/btn_nav_profile" />
+
     </group>
 </menu>
