+++ app/src/main/res/values/strings.xml
@@ -1,0 +1,1 @@
+<resources></resources>
